RNA working silution plates use different purpose
Fluidigm submissions in Sequencescape are made from working dilution
plates. These are selected using a purpose and a sample name. However
if a sample has come across the bridge, but DNA and RNA samples are
on Working Dilution plates, and Sequencescape can't distinguish between
them. (Except by barcode prefix, but using that would break pattern)

I've added a new purpose to Sequencescape for RNA Working Diltion.
I believe these changes are all thats needed to ensure the bridge uses it,
but as I'm not familiar with the inner workings of S2 or the bridge,
I may have missed something important.